Implement abstract class and methods for Playlist functionality
